requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Cors>=4.0.0
Werkzeug>=3.0.0
SQLAlchemy>=2.0.0
PyJWT>=2.7.0
requests>=2.28.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
